refactor(d2): extract context module + add CLI context command
Move get_project_context(), get_belief_metrics(), read_patterns(),
extract_summary() from server.rs to src/commands/context.rs. MCP server
imports from the extracted module. server.rs drops ~260 lines.

Add `patina context [--topic]` CLI command — thin wrapper over the
shared module. CLI-connected LLMs can now access patterns and beliefs.

Update D2 spec: revised problem statement (D1 shipped, context is
broken), reframed design, 5-commit implementation plan, new exit
criteria, Layer 3 deferred to stretch.

Zero behavior change — MCP and CLI return identical output.
